Extract duplicate test helpers into shared scope
