+++ EssentialFeed/EssentialFeedTests/FeedAPI/URLSessionHTTPClientTests.swift
@@ -88,11 +88,7 @@ class URLSessionHTTPClientTests: XCTestCase {
         trackForMemoryLeaks(sut, file: file, line: line)
         return sut
     }
-    
-    private func anyURL() -> URL {
-        return URL(string: "https://any-url.com")!
-    }
-    
+        
     private func nonHTTPURLResponse() -> URLResponse {
         return URLResponse(url: anyURL(), mimeType: nil, expectedContentLength: 0, textEncodingName: nil)
     }
@@ -104,11 +100,7 @@ class URLSessionHTTPClientTests: XCTestCase {
     private func anyData() -> Data {
         return Data("any".utf8)
     }
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "Invalid request", code: 400)
-    }
-    
+        
     private func resultErrorOn(data: Data?, response: URLResponse?, error: Error?, file: StaticString = #filePath, line: UInt = #line) -> Error? {
         let receivedResult = resultFor(data: data, response: response, error: error, file: file, line: line)
