+++ EssentialFeed/EssentialFeedTests/FeedCache/CacheFeedUseCaseTests.swift
@@ -124,22 +124,5 @@ class CacheFeedUseCaseTests: XCTestCase {
 
         wait(for: [exp], timeout: 1.0)
         XCTAssertEqual(receivedError as? NSError, expectedError, file: file, line: line)
-    }
-    
-    private func uniqueImage() -> FeedImage {
-        return .init(id: UUID(), description: nil, location: nil, url: URL(string: "https://a-url.com")!)
-    }
-    
-    private func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
-        let feed = [uniqueImage(), uniqueImage()]
-        let localFeed = feed.map { LocalFeedImage(id: $0.id,
-                                                   description: $0.description,
-                                                   location: $0.location,
-                                                   url: $0.url)}
-        return (feed, localFeed)
-    }
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "Invalid request", code: 400)
-    }
+    }    
 }
