+++ EssentialFeed/EssentialFeedTests/FeedCache/ValidateFeedCacheUseCaseTests.swift
@@ -56,31 +56,4 @@ class ValidateFeedCacheUseCaseTests: XCTestCase {
         
         return (sut, store)
     }
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "Invalid request", code: 400)
-    }
-    
-    private func uniqueImage() -> FeedImage {
-        return .init(id: UUID(), description: nil, location: nil, url: URL(string: "https://a-url.com")!)
-    }
-
-    private func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
-        let feed = [uniqueImage(), uniqueImage()]
-        let localFeed = feed.map { LocalFeedImage(id: $0.id,
-                                                   description: $0.description,
-                                                   location: $0.location,
-                                                   url: $0.url)}
-        return (feed, localFeed)
-    }
-}
-
-private extension Date {
-    func addingDays(_ days: Int) -> Self {
-        return Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
-    }
-    
-    func addingSeconds(_ seconds: TimeInterval) -> Self {
-        return self + seconds
-    }
 }
